C-10 [*] переименование скрипта поиска ошибок
